requirements.txt
fastapi==0.115.0
uvicorn==0.30.1

python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
supabase
email-validator
python-docx
PyPDF2
PyMuPDF
pytesseract
fuzzywuzzy
python-Levenshtein
scikit-learn
requests
rapidfuzz
